Move book market money utils to general utils
